# backend/internal/model/schema.go
// Package model provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.9.1 DO NOT EDIT.
package model

import (
	"time"
)

const (
	BearerAuthScopes = "BearerAuth.Scopes"
)

// Defines values for CommonResponseStatus.
const (
	CommonResponseStatusError CommonResponseStatus = "error"

	CommonResponseStatusOk CommonResponseStatus = "ok"
)

// CommonResponse defines model for CommonResponse.
type CommonResponse struct {
	Message string               `json:"message"`
	Status  CommonResponseStatus `json:"status"`
}

// CommonResponseStatus defines model for CommonResponse.Status.
type CommonResponseStatus string

// Merchant defines model for Merchant.
type Merchant struct {
	Code      string     `json:"code"`
	CreatedAt *time.Time `json:"created_at,omitempty"`
	Id        string     `json:"id"`
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
}

// NewMerchant defines model for NewMerchant.
type NewMerchant struct {
	Code string `json:"code"`
}

// NewTeamMember defines model for NewTeamMember.
type NewTeamMember struct {
	Email      string `json:"email"`
	MerchantId string `json:"merchant_id"`
}

// TeamMember defines model for TeamMember.
type TeamMember struct {
	CreatedAt  *time.Time `json:"created_at,omitempty"`
	Email      string     `json:"email"`
	Id         string     `json:"id"`
	MerchantId string     `json:"merchant_id"`
	UpdatedAt  *time.Time `json:"updated_at,omitempty"`
}

// UpdateMerchant defines model for UpdateMerchant.
type UpdateMerchant struct {
	Code string `json:"code"`
}

// UpdateTeamMember defines model for UpdateTeamMember.
type UpdateTeamMember struct {
	Email      string `json:"email"`
	MerchantId string `json:"merchant_id"`
}

// PageLimit defines model for PageLimit.
type PageLimit int

// PageOffset defines model for PageOffset.
type PageOffset int

// N400Error defines model for 400Error.
type N400Error CommonResponse

// N404Error defines model for 404Error.
type N404Error CommonResponse

// N409Error defines model for 409Error.
type N409Error CommonResponse

// N500Error defines model for 500Error.
type N500Error CommonResponse

// GetMerchantParams defines parameters for GetMerchant.
type GetMerchantParams struct {
	// Limits the number of items on a page
	Limit PageLimit `json:"limit"`

	// Specifies the page number of the item to be displayed
	Offset PageOffset `json:"offset"`
}

// PostMerchantJSONBody defines parameters for PostMerchant.
type PostMerchantJSONBody NewMerchant

// PutMerchantIdJSONBody defines parameters for PutMerchantId.
type PutMerchantIdJSONBody UpdateMerchant

// PostTeammemberJSONBody defines parameters for PostTeammember.
type PostTeammemberJSONBody NewTeamMember

// GetTeammemberMerchantMerchantIdParams defines parameters for GetTeammemberMerchantMerchantId.
type GetTeammemberMerchantMerchantIdParams struct {
	// Limits the number of items on a page
	Limit PageLimit `json:"limit"`

	// Specifies the page number of the item to be displayed
	Offset PageOffset `json:"offset"`
}

// PutTeammemberIdJSONBody defines parameters for PutTeammemberId.
type PutTeammemberIdJSONBody UpdateTeamMember

// PostMerchantJSONRequestBody defines body for PostMerchant for application/json ContentType.
type PostMerchantJSONRequestBody PostMerchantJSONBody

// PutMerchantIdJSONRequestBody defines body for PutMerchantId for application/json ContentType.
type PutMerchantIdJSONRequestBody PutMerchantIdJSONBody

// PostTeammemberJSONRequestBody defines body for PostTeammember for application/json ContentType.
type PostTeammemberJSONRequestBody PostTeammemberJSONBody

// PutTeammemberIdJSONRequestBody defines body for PutTeammemberId for application/json ContentType.
type PutTeammemberIdJSONRequestBody PutTeammemberIdJSONBody